+++ packages/dumi-theme-vant/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.1.0 (2022-05-18)
+
+
+### Features
+
+* add dark theme ([ee83aff](https://github.com/bijinfeng/rn-vant/commit/ee83aff6280d0f3604f89156ce3f68e5bf4e2b4a))
+* add dumi-theme-vant ([7fd03c6](https://github.com/bijinfeng/rn-vant/commit/7fd03c6659b580d24f12a6bc0dd7f18336dc6afa))
+* add SourceCode ([2960ef7](https://github.com/bijinfeng/rn-vant/commit/2960ef7c7d0fc25c29badf1befcefc1cd3735a97))
+* add Typography component ([08fbb1d](https://github.com/bijinfeng/rn-vant/commit/08fbb1d30d72133163d200263be19a465a7be848))
+* change package name ([2851a86](https://github.com/bijinfeng/rn-vant/commit/2851a864a0842a035b3f9affd1deb92496071a4c))
+* update dark theme ([e728c50](https://github.com/bijinfeng/rn-vant/commit/e728c5030a4d3b56c465d7740b6bc893decbfbd0))
+
+
+
+
+
 ## [0.0.2](https://github.com/bijinfeng/rn-vant/compare/@rn-vant/dumi-theme-vant@0.1.0...@rn-vant/dumi-theme-vant@0.0.2) (2022-05-14)
 
 **Note:** Version bump only for package @rn-vant/dumi-theme-vant