+++ packages/rn-vant-icons/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.1.0 (2022-05-18)
+
+
+### Features
+
+* change package name ([2851a86](https://github.com/bijinfeng/rn-vant/commit/2851a864a0842a035b3f9affd1deb92496071a4c))
+
+
+
+
+
 ## [0.0.2](https://github.com/bijinfeng/rn-vant/compare/@rn-vant/icons@0.1.0...@rn-vant/icons@0.0.2) (2022-05-14)
 
 **Note:** Version bump only for package @rn-vant/icons