+++ packages/rn-vant/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.1.0 (2022-05-18)
+
+
+### Features
+
+* add Typography component ([08fbb1d](https://github.com/bijinfeng/rn-vant/commit/08fbb1d30d72133163d200263be19a465a7be848))
+* change package name ([2851a86](https://github.com/bijinfeng/rn-vant/commit/2851a864a0842a035b3f9affd1deb92496071a4c))
+
+
+
+
+
 ## [0.0.2](https://github.com/bijinfeng/rn-vant/compare/rn-vant@0.1.0...rn-vant@0.0.2) (2022-05-14)
 
 **Note:** Version bump only for package rn-vant